Use io.ReadCloser + json.NewDecoder
In order to use json.NewDecoder, we have to operate on an I/O stream.

So, we swap out passing around a byteslice for this instead.

This means we can't defer closure of the body stream within the
`doRawRequest` method and have to unwrap it into individual close
statements.

But we can in the higher level `doRequest` method.

Finally, in `Repository.GetFileBlob`, we utilize `ioutil.ReadAll` to
replicate the original functionality to consume the response body.