Create RepositoryRefsOptions and ListRefs method with helper funds (#155)
* Create Rep Ref Options type

* Create list refs functions

This commit creates a ListRefs method of the Repository type.
Specifically, it creates a RepositoryRefs struct type that is returned
from the decoodeRepositoryRefs function that is a helper function
for decoding the reponse from the refs GET API Bitbucket endpoint.

* Correct Options type and add documentation

* Create test for ListRefs

This commit creates a test case for ListRefs based on setting up the
test repo with a new branch and then listing the refs and making sure
properly created the test repo. So, this test case technically also
tests a portion of the functionality of the CreateBranch function.