ISSUE #166 Update GetDiffStat functionality
This commit updates the GetDiffStat functionality so that now users
can add in a slice of strings named Fields in the DiffStatOptions.
Using this option users can exclude or only include certain parts
of the JSON response.

Note, currently, given how the json decoding is although we remove the
specified fields from the JSON response, they get unmarshalled as zero
values for the corresponding DiffStatRes fields.

Also, a test was added, but we should probably improve it by testing
whether or not what we removed/selected is present accordingly.